test(ui): Move shared component tests out of source files
Move the `PluginMultiSelectField` helper functions to
`plugin-multi-select-field-utils.ts` so the external test can import them.

Exporting those helpers directly from `plugin-multi-select-field.tsx` would
violate the React Fast Refresh lint rule that component files only export
components.
